+++ sktime/_dists_kernels_base/__init__.py
@@ -1,0 +1,13 @@
+# -*- coding: utf-8 -*-
+"""Module exports for dist_kernels module."""
+
+from sktime._dists_kernels_base._base import (
+    BasePairwiseTransformer,
+    BasePairwiseTransformerPanel,
+)
+
+# TODO: remove dir in 0.15.0
+__all__ = [
+    "BasePairwiseTransformer",
+    "BasePairwiseTransformerPanel",
+]
