+++ sktime/dists_kernels/distances/__init__.py
@@ -1,20 +1,12 @@
 # -*- coding: utf-8 -*-
 """Module exports for dist_kernels module."""
 
-from sktime.dists_kernels._base import (
-    BasePairwiseTransformer,
-    BasePairwiseTransformerPanel,
-)
-
-# todo remove these imports in 0.15.0
 from sktime.dists_kernels.compose_tab_to_panel import AggrDist, FlatDist
 from sktime.dists_kernels.dtw import DtwDist
 from sktime.dists_kernels.edit_dist import EditDist
 from sktime.dists_kernels.scipy_dist import ScipyDist
 
 __all__ = [
-    "BasePairwiseTransformer",
-    "BasePairwiseTransformerPanel",
     "AggrDist",
     "DtwDist",
     "EditDist",
