+++ sktime/registry/_base_classes.py
@@ -58,16 +58,16 @@ __author__ = ["fkiraly"]
 
 import pandas as pd
 
+from sktime._dists_kernels_base._base import (
+    BasePairwiseTransformer,
+    BasePairwiseTransformerPanel,
+)
 from sktime.alignment.base import BaseAligner
 from sktime.annotation.base import BaseSeriesAnnotator
 from sktime.base import BaseEstimator, BaseObject
 from sktime.classification.base import BaseClassifier
 from sktime.classification.early_classification_base import BaseEarlyClassifier
 from sktime.clustering.base import BaseClusterer
-from sktime.dists_kernels._base import (
-    BasePairwiseTransformer,
-    BasePairwiseTransformerPanel,
-)
 from sktime.forecasting.base import BaseForecaster
 from sktime.forecasting.model_selection._split import BaseSplitter
 from sktime.networks.base import BaseDeepNetwork
